clean up python impls in raphtory-api
